+++ attributes/default.rb
@@ -19,3 +19,4 @@
 #
 
 default['dnsmasq_local']['config'] = {}
+default['dnsmasq_local']['environment'] = {}