+++ libraries/resource_dnsmasq_local_service.rb
@@ -25,10 +25,53 @@ class Chef
     # A Chef resource for acting on the dnsmasq service.
     #
     # @author Jonathan Hartman <jonathan.hartman@socrata.com>
-    class DnsmasqLocalService < LWRPBase
+    class DnsmasqLocalService < LWRPBase # rubocop:disable Style/Documentation
       self.resource_name = :dnsmasq_local_service
-      actions Chef::Resource::Service.new('_', nil).allowed_actions
-      default_action [:enable, :start]
+      actions Chef::Resource::Service.new('_', nil).allowed_actions + \
+              [:create, :remove].uniq
+      default_action [:create, :enable, :start]
+
+      #
+      # Set up a default hash of environment variables that a user can
+      # override in its entirety by passing in a new one. Every key set will be
+      # upcased and written out to `/etc/default/dnsmasq` on Ubuntu nodes.
+      #
+      attribute :environment,
+                kind_of: Hash,
+                default: {
+                  enabled: 1,
+                  config_dir: '/etc/dnsmasq.d,.dpkg-dist,.dpkg-old,.dpkg-new'
+                }
+
+      #
+      # Allow individual attributes to be fed in and merged with the default
+      # environment without blowing away the entire thing.
+      #
+      # (see Chef::Resource#method_missing)
+      #
+      def method_missing(method_symbol, *args, &block)
+        if block.nil? && args.length == 1
+          self.class.attribute method_symbol, kind_of: args[0].class
+          add_state_attr(method_symbol)
+          send(method_symbol, args[0]) unless args[0].nil?
+        else
+          super
+        end
+      end
+
+      #
+      # Add a newly-declared attribute into the list of desired state
+      # attributes to account for the fact that Chef 12 defaults this to true
+      # while 11 defaults it to false.
+      #
+      # @param attr [Symbol] the new attribute to add to the list of state ones
+      #
+      def add_state_attr(attr)
+        new_attrs = (self.class.state_attrs << attr).uniq
+        self.class.state_attrs(*new_attrs)
+      end
     end
   end
-end
+end unless defined?(Chef::Resource::DnsmasqLocalService)
+# Don't let this class be reloaded or strange things happen to the custom
+# attributes we've loaded in via `method_missing`.