+++ spec/resources/dnsmasq_local/ubuntu/14_04_spec.rb
@@ -2,12 +2,12 @@ require_relative '../../../spec_helper'
 
 describe 'resource_dnsmasq_local::ubuntu::14_04' do
   let(:name) { 'default' }
-  %i(config action).each { |p| let(p) { nil } }
+  %i(config environment action).each { |p| let(p) { nil } }
   let(:runner) do
     ChefSpec::SoloRunner.new(
       step_into: 'dnsmasq_local', platform: 'ubuntu', version: '14.04'
     ) do |node|
-      %i(name config action).each do |p|
+      %i(name config environment action).each do |p|
         node.set['resource_dnsmasq_local_test'][p] = send(p) unless send(p).nil?
       end
     end
@@ -41,7 +41,15 @@ describe 'resource_dnsmasq_local::ubuntu::14_04' do
           .to notify('dnsmasq_local_service[default]').to(:restart)
       end
 
-      it 'enables and starts the dnsmasq_local_service' do
+      it 'creates, enables, and starts the dnsmasq_local_service' do
+        expected = {
+          environment: {
+            config_dir: '/etc/dnsmasq.d,.dpkg-dist,.dpkg-old,.dpkg-new',
+            enabled: 1
+          }
+        }.merge(environment.to_h)
+        expect(chef_run).to create_dnsmasq_local_service('default')
+          .with(expected)
         expect(chef_run).to enable_dnsmasq_local_service('default')
         expect(chef_run).to start_dnsmasq_local_service('default')
       end
@@ -60,15 +68,23 @@ describe 'resource_dnsmasq_local::ubuntu::14_04' do
 
       it_behaves_like 'any attribute set'
     end
+
+    context 'an environment attribute' do
+      let(:environment) { { example: 'elpmaxe' } }
+      cached(:chef_run) { converge }
+
+      it_behaves_like 'any attribute set'
+    end
   end
 
   context 'the :remove action' do
     let(:action) { :remove }
     cached(:chef_run) { converge }
 
-    it 'stops and disables the dnsmasq_local_service' do
+    it 'stops, disables, and removes the dnsmasq_local_service' do
       expect(chef_run).to stop_dnsmasq_local_service('default')
       expect(chef_run).to disable_dnsmasq_local_service('default')
+      expect(chef_run).to remove_dnsmasq_local_service('default')
     end
 
     it 'removes the dnsmasq_local_config' do