+++ spec/support/cookbooks/resource_dnsmasq_local_test/recipes/default.rb
@@ -4,5 +4,6 @@ attrs = node['resource_dnsmasq_local_test']
 
 dnsmasq_local attrs['name'] do
   config attrs['config'] unless attrs['config'].nil?
+  environment attrs['environment'] unless attrs['environment'].nil?
   action attrs['action'] unless attrs['action'].nil?
 end